requirements.txt
PyYAML==5.4.1
requests==2.26.0
tabulate==0.8.9
webexteamssdk==1.6
fastapi==0.85.1
uvicorn==0.18.3